+++ easygraphics/processing/processing.py
@@ -39,6 +39,7 @@ __all__ = [
     'bezier', 'draw_bezier', 'lines', 'draw_lines', 'poly_line', 'draw_poly_line', 'polygon', 'draw_polygon',
     'fill_polygon', 'rect', 'draw_rect', 'fill_rect', 'rounded_rect', 'draw_rounded_rect', 'fill_rounded_rect',
     'flood_fill', 'draw_image', 'clear', 'clear_view_port',
+    'quadratic', 'draw_quadratic', 'begin_shape', 'end_shape', 'vertex', 'bezier_vertex', 'quadratic_vertex',
     # text functions #
     'draw_text', 'draw_rect_text', 'text_width', 'text_height',
     # time control functions#
@@ -349,12 +350,40 @@ def fill_chord(x: float, y: float, start_angle: float, end_angle: float, radius_
     _widget.get_canvas().fill_chord(x, y, start_angle, end_angle, radius_x, radius_y)
 
 
-def bezier(control_points: List[float]):
-    _widget.get_canvas().bezier(control_points)
+def begin_shape():
+    _widget.get_canvas().begin_shape()
 
 
-def draw_bezier(control_points: List[float]):
-    _widget.get_canvas().draw_bezier(control_points)
+def end_shape():
+    _widget.get_canvas().end_shape()
+
+
+def vertex(x: float, y: float):
+    _widget.get_canvas().vertex(x, y)
+
+
+def bezier_vertex(x1: float, y1: float, x2: float, y2: float, x3: float, y3: float):
+    _widget.get_canvas().bezier_vertex(x1, y1, x2, y2, x3, y3)
+
+
+def quadratic_vertex(x1: float, y1: float, x2: float, y2: float):
+    _widget.get_canvas().quadratic_vertex(x1, y1, x2, y2)
+
+
+def bezier(x0: float, y0: float, x1: float, y1: float, x2: float, y2: float, x3: float, y3: float):
+    _widget.get_canvas().bezier(x0, y0, x1, y1, x2, y2, x3, y3)
+
+
+def draw_bezier(x0: float, y0: float, x1: float, y1: float, x2: float, y2: float, x3: float, y3: float):
+    _widget.get_canvas().draw_bezier(x0, y0, x1, y1, x2, y2, x3, y3)
+
+
+def quadratic(x0: float, y0: float, x1: float, y1: float, x2: float, y2: float):
+    _widget.get_canvas().bezier(x0, y0, x1, y1, x2, y2)
+
+
+def draw_quadratic(x0: float, y0: float, x1: float, y1: float, x2: float, y2: float):
+    _widget.get_canvas().draw_bezier(x0, y0, x1, y1, x2, y2)
 
 
 def lines(points: List[float]):
@@ -494,13 +523,23 @@ def get_canvas() -> Image:
     return _widget.get_canvas()
 
 
-def run_app(draw_func, setup_func=None):
-    global _app, _widget, setup, draw
-    if setup_func is not None:
-        setup = setup_func
-    if draw_func is None:
-        raise RuntimeError("must provide draw() function!")
-    draw = draw_func
+def run_app(_globals):
+    global _app, _widget, setup, draw, on_mouse_clicked
+    global on_mouse_pressed, on_mouse_released, on_mouse_dragged, on_mouse_wheel
+    if 'setup' in _globals:
+        setup = _globals['setup']
+    if 'draw' in _globals:
+        draw = _globals['draw']
+    if 'on_mouse_clicked' in _globals:
+        on_mouse_clicked = _globals['on_mouse_clicked']
+    if 'on_mouse_pressed' in _globals:
+        on_mouse_pressed = _globals['on_mouse_pressed']
+    if 'on_mouse_released' in _globals:
+        on_mouse_released = _globals['on_mouse_released']
+    if 'on_mouse_dragged' in _globals:
+        on_mouse_dragged = _globals['on_mouse_dragged']
+    if 'on_mouse_wheel' in _globals:
+        on_mouse_wheel = _globals['on_mouse_wheel']
     _app = QtWidgets.QApplication([])
     _widget = _Processing_Widget()
     _widget.init()