+++ examples/tutorials/04_bezier.py
@@ -6,6 +6,6 @@ if __name__ == "__main__":
 
     init_graph(600, 400)
     points = [300, 50, 200, 50, 200, 200, 100, 200]
-    draw_bezier(points)
+    draw_bezier(*points)
     pause()
     close_graph()